chore(SRC): remove some circular deps, move shared consts
