+++ src/lib/containers/Search.tsx
@@ -13,7 +13,7 @@ import TotalQueryResults from './TotalQueryResults'
 import getColorPallette from './ColorGradient'
 import { SEARCH_CLASS_CSS } from './QueryWrapperMenu'
 import { Dropdown } from 'react-bootstrap'
-import { unCamelCase } from './table/SynapseTable'
+import { unCamelCase } from '../utils/UtilityFns'
 
 library.add(faCaretDown)
 library.add(faCaretUp)
