+++ src/lib/containers/download_list/index.ts
@@ -1,7 +1,6 @@
 import DownloadDetails from './DownloadDetails'
 import {DownloadConfirmation} from './DownloadConfirmation'
-import './DownloadList.scss'
-import './DownloadConfirmation.scss'
+
 
 export default { DownloadDetails } 
 export { DownloadDetails, DownloadConfirmation}
