+++ src/lib/containers/table/SynapseTableConstants.ts
@@ -1,0 +1,6 @@
+export const TOOLTIP_DELAY_SHOW = 500
+export const SELECT_ALL = 'SELECT_ALL';
+export const DESELECT_ALL = 'DESELECT_ALL';
+export const DOWNLOAD_FILES_MENU_TEXT = 'Download Files'; // double check these icons!
+export const ICON_STATE: string[] = ['sort-amount-down', 'sort-amount-down', 'sort-amount-up'];
+
