+++ src/lib/containers/table/table-top/ColumnSelection.tsx
@@ -4,7 +4,7 @@ import { SelectColumn } from '../../../utils/jsonResponses/Table/SelectColumn'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import ColumnsSvg from '../../../assets/icons/columns.svg'
 import ReactTooltip from 'react-tooltip'
-import { TOOLTIP_DELAY_SHOW } from '../SynapseTable'
+import { TOOLTIP_DELAY_SHOW } from '../SynapseTableConstants'
 
 type ColumnSelectionProps = {
   headers: SelectColumn[]
